# docs/swagger.yaml
basePath: /
definitions:
...
  entity.UserInfoRes:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: integer
    type: object
...
  user.emailData:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  user.loginData:
    properties:
      account:
        type: string
      password:
        type: string
    required:
    - account
    - password
    type: object
  user.loginRes:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: integer
      token:
        type: string
    type: object
  user.loginTicketData:
    properties:
      authorization:
        type: string
      url:
        type: string
    required:
    - authorization
    - url
    type: object
  user.nameData:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  user.passwordData:
    properties:
      new:
        type: string
      old:
        type: string
    required:
    - new
    - old
    type: object
  user.registerData:
    properties:
      account:
        type: string
      password:
        type: string
    required:
    - account
    - password
    type: object
info:
  contact:
    name: Mogg
    url: https://github.com/MarcusMogg
  description: buaashow is a sample RESTful api server.
  title: buaashow
  version: "1.0"
paths:
...
  /user/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新邮箱
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/user.emailData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.loginRes'
      summary: 修改邮箱, 需用户登录
      tags:
      - User
...
  /user/info/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserInfoRes'
      summary: 获取指定id(account)的用户信息
      tags:
      - User
  /user/infolist:
    get:
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: account target
        in: query
        name: account
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserInfoRes'
            type: array
      summary: 获取用户列表，only ADMIN
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号密码
        in: body
        name: logindata
        required: true
        schema:
          $ref: '#/definitions/user.loginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.loginRes'
      summary: 使用账号密码登录
      tags:
      - User
  /user/name:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新name
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/user.nameData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.loginRes'
      summary: 修改name, 需用户登录
      tags:
      - User
  /user/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新旧密码
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/user.passwordData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.loginRes'
      summary: 修改密码, 需用户登录
      tags:
      - User
  /user/teacher:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号密码必选，邮箱可选
        in: body
        name: logindata
        required: true
        schema:
          $ref: '#/definitions/user.registerData'
      produces:
      - application/json
      summary: 创建教师账号 only ADMIN
      tags:
      - User
  /user/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 云平台返回的ticket
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/user.loginTicketData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.loginRes'
      summary: 使用云平台登录
      tags:
      - User
swagger: "2.0"